gcloudcomputeengine/requirements.txt
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-cloud-compute>=1.0.0
python-dotenv>=0.19.0
google-api-core>=2.0.0
requests>=2.31.0